+++ polars/polars-lazy/src/physical_plan/expressions/alias.rs
@@ -1,0 +1,63 @@
+use crate::physical_plan::state::ExecutionState;
+use crate::prelude::*;
+use polars_core::frame::groupby::GroupTuples;
+use polars_core::prelude::*;
+use std::sync::Arc;
+
+pub struct AliasExpr {
+    pub(crate) physical_expr: Arc<dyn PhysicalExpr>,
+    pub(crate) name: Arc<String>,
+    expr: Expr,
+}
+
+impl AliasExpr {
+    pub fn new(physical_expr: Arc<dyn PhysicalExpr>, name: Arc<String>, expr: Expr) -> Self {
+        Self {
+            physical_expr,
+            name,
+            expr,
+        }
+    }
+}
+
+impl PhysicalExpr for AliasExpr {
+    fn as_expression(&self) -> &Expr {
+        &self.expr
+    }
+
+    fn evaluate(&self, df: &DataFrame, state: &ExecutionState) -> Result<Series> {
+        let mut series = self.physical_expr.evaluate(df, state)?;
+        series.rename(&self.name);
+        Ok(series)
+    }
+
+    fn to_field(&self, input_schema: &Schema) -> Result<Field> {
+        Ok(Field::new(
+            &self.name,
+            self.physical_expr
+                .to_field(input_schema)?
+                .data_type()
+                .clone(),
+        ))
+    }
+
+    fn as_agg_expr(&self) -> Result<&dyn PhysicalAggregation> {
+        Ok(self)
+    }
+}
+
+impl PhysicalAggregation for AliasExpr {
+    fn aggregate(
+        &self,
+        df: &DataFrame,
+        groups: &GroupTuples,
+        state: &ExecutionState,
+    ) -> Result<Option<Series>> {
+        let agg_expr = self.physical_expr.as_agg_expr()?;
+        let opt_agg = agg_expr.aggregate(df, groups, state)?;
+        Ok(opt_agg.map(|mut agg| {
+            agg.rename(&self.name);
+            agg
+        }))
+    }
+}
