+++ polars/polars-lazy/src/physical_plan/expressions/default.rs
@@ -9,48 +9,6 @@ use rayon::prelude::*;
 use std::borrow::Cow;
 use std::sync::Arc;
 
-pub struct AliasExpr {
-    pub(crate) physical_expr: Arc<dyn PhysicalExpr>,
-    pub(crate) name: Arc<String>,
-    expr: Expr,
-}
-
-impl AliasExpr {
-    pub fn new(physical_expr: Arc<dyn PhysicalExpr>, name: Arc<String>, expr: Expr) -> Self {
-        Self {
-            physical_expr,
-            name,
-            expr,
-        }
-    }
-}
-
-impl PhysicalExpr for AliasExpr {
-    fn as_expression(&self) -> &Expr {
-        &self.expr
-    }
-
-    fn evaluate(&self, df: &DataFrame, state: &ExecutionState) -> Result<Series> {
-        let mut series = self.physical_expr.evaluate(df, state)?;
-        series.rename(&self.name);
-        Ok(series)
-    }
-
-    fn to_field(&self, input_schema: &Schema) -> Result<Field> {
-        Ok(Field::new(
-            &self.name,
-            self.physical_expr
-                .to_field(input_schema)?
-                .data_type()
-                .clone(),
-        ))
-    }
-
-    fn as_agg_expr(&self) -> Result<&dyn PhysicalAggregation> {
-        Ok(self)
-    }
-}
-
 pub struct IsNullExpr {
     physical_expr: Arc<dyn PhysicalExpr>,
     expr: Expr,
