+++ polars/polars-lazy/src/physical_plan/expressions/final_agg.rs
@@ -11,22 +11,6 @@ use polars_core::frame::groupby::{fmt_groupby_column, GroupByMethod, GroupTuples
 use polars_core::prelude::*;
 use polars_core::utils::NoNull;
 
-impl PhysicalAggregation for AliasExpr {
-    fn aggregate(
-        &self,
-        df: &DataFrame,
-        groups: &GroupTuples,
-        state: &ExecutionState,
-    ) -> Result<Option<Series>> {
-        let agg_expr = self.physical_expr.as_agg_expr()?;
-        let opt_agg = agg_expr.aggregate(df, groups, state)?;
-        Ok(opt_agg.map(|mut agg| {
-            agg.rename(&self.name);
-            agg
-        }))
-    }
-}
-
 fn rename_option_series(opt: Option<Series>, name: &str) -> Option<Series> {
     opt.map(|mut s| {
         s.rename(name);
